Separation and modification
Separated the Depth camera access class to a diferent file, removed the HandTrackingModule.py since it could only track one hand at a time.
Made small esthetic changes to the main method.
